+++ src/main/java/animores/serverapi/pet/service/impl/PetServiceImpl.java
@@ -1,6 +1,8 @@
 package animores.serverapi.pet.service.impl;
 
 import animores.serverapi.account.domain.Account;
+import animores.serverapi.common.exception.CustomException;
+import animores.serverapi.common.exception.ExceptionCode;
 import animores.serverapi.pet.dao.PetDao;
 import animores.serverapi.pet.entity.Breed;
 import animores.serverapi.pet.entity.Pet;
@@ -18,7 +20,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @RequiredArgsConstructor
 @Service
@@ -26,6 +30,25 @@ public class PetServiceImpl implements PetService {
     private final SpeciesRepository speciesRepository;
     private final BreedRepository breedRepository;
     private final PetRepository petRepository;
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<Pet> checkAccountPets(Long accountId, List<Long> petIds) {
+        List<Pet> pets = petRepository.findAllByAccount_id(accountId);
+
+        if(petIds.isEmpty()){
+            return pets;
+        } else {
+            Set<Long> petIdsSet = new HashSet<>(petIds);
+            pets.forEach(pet -> {
+                if(!petIdsSet.contains(pet.getId())){
+                    throw new CustomException(ExceptionCode.ILLEGAL_PET_IDS);
+                }
+            });
+
+            return pets.stream().filter(pet -> petIdsSet.contains(pet.getId())).toList();
+        }
+    }
 
     @Override
     public List<SpeciesResponse> getSpecies() {
